Rename slang-wgsl to slang-wasm
