+++ index.html
@@ -1,6 +1,6 @@
 <html>
   <script src="try-slang.js"> </script>
-  <script src="slang-wgsl.js"></script>
+  <script src="slang-wasm.js"></script>
   <head>
     <title>Try Slang!</title>
   </head>
