+++ slang-wasm.cpp
@@ -4,7 +4,7 @@
 #include <slang.h>
 #include <slang-com-ptr.h>
 #include "try-slang-external-includes-end.h"
-#include "slang-wgsl.h"
+#include "slang-wasm.h"
 
 using namespace slang;
